+++ src/main/java/spoilagesystem/listeners/EntityDeathListener.java
@@ -1,6 +1,5 @@
 package spoilagesystem.listeners;
 
-import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -24,7 +23,7 @@ public final class EntityDeathListener implements Listener {
         if (event.getEntity() instanceof Player) return;
         List<ItemStack> newDrops = new ArrayList<>();
         for (ItemStack drop : event.getDrops()) {
-            if (drop.getType().isEdible() && drop.getType() != Material.ROTTEN_FLESH
+            if (drop.getType().isEdible() && !timeStampService.isSpoiledFoodMaterial(drop)
                     && !timeStampService.isWaxed(drop)) {
                 newDrops.add(timeStampService.assignTimeStamp(drop));
             } else {