+++ src/main/java/spoilagesystem/listeners/PlayerJoinListener.java
@@ -1,6 +1,5 @@
 package spoilagesystem.listeners;
 
-import org.bukkit.Material;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
@@ -21,7 +20,7 @@ public final class PlayerJoinListener implements Listener {
     public void onPlayerJoin(PlayerJoinEvent event) {
         Arrays.stream(event.getPlayer().getInventory().getContents())
                 .filter(Objects::nonNull)
-                .filter(item -> item.getType().isEdible() && item.getType() != Material.ROTTEN_FLESH
+                .filter(item -> item.getType().isEdible() && !timeStampService.isSpoiledFoodMaterial(item)
                         && !timeStampService.isWaxed(item))
                 .forEach(item -> {
                     if (!timeStampService.timeStampAssigned(item)) {